requirements.txt
discord.py[voice]
oauth2client
python-dotenv
gspread
